Refactor and fix top level flag handling
The vars designed to hold the global flag values were being used to
create the command before the associated flag values were read. This
patch moves the vars and does the flag parsing to a place where the
vars will be in the right scope. I believe this bug was introduced
in PR #1021.

Resolves: https://issues.redhat.com/browse/HACBS-2688
